modularising code
